Changed species integers to bytes.

+++ src/PokeGame.Infrastructure/Converters/CatchRateConverter.cs
@@ -6,7 +6,7 @@ internal class CatchRateConverter : JsonConverter<CatchRate>
 {
   public override CatchRate? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
   {
-    return reader.TryGetInt32(out int value) ? new CatchRate(value) : null;
+    return reader.TryGetByte(out byte value) ? new CatchRate(value) : null;
   }
 
   public override void Write(Utf8JsonWriter writer, CatchRate catchRate, JsonSerializerOptions options)
